first version of pdf list with barcodes
