+++ plugins/Admin/src/Template/Element/pdf/invoice.ctp
@@ -13,12 +13,12 @@
  * @link          https://www.foodcoopshop.com
  */
 
-use App\Lib\Pdf\AppTcpdf;
+use App\Lib\Pdf\ListTcpdf;
 use Cake\Core\Configure;
 use Cake\ORM\TableRegistry;
 use Cake\Filesystem\Folder;
 
-$pdf = new AppTcpdf();
+$pdf = new ListTcpdf();
 $pdf->setTextHelper($this->Text);
 $pdf->SetLeftMargin(12);
 $pdf->AddPage();
