+++ src/Template/pdf/generate_right_of_withdrawal_information_and_form.ctp
@@ -14,11 +14,11 @@
  */
 
 use App\Controller\Component\StringComponent;
-use App\Lib\Pdf\AppTcpdf;
+use App\Lib\Pdf\ListTcpdf;
 use Cake\Core\Configure;
 use Cake\I18n\I18n;
 
-$pdf = new AppTcpdf();
+$pdf = new ListTcpdf();
 $pdf->SetLeftMargin(12);
 $pdf->SetRightMargin(12);
